+++ packages/sampler/stories/qa-stories/Scroller.js
@@ -1,0 +1,33 @@
+import React from 'react';
+import ExpandableList from '@enact/moonstone/ExpandableList';
+import Selectable from '@enact/ui/Selectable';
+import Scroller from '@enact/moonstone/Scroller';
+
+const StatefullExpandableList = Selectable(ExpandableList);
+
+const itemData = [];
+for (let i = 0; i < 100; i++) {
+	itemData.push(`Item ${i}`);
+}
+
+import {storiesOf} from '@kadira/storybook';
+import {withKnobs} from '@kadira/storybook-addon-knobs';
+
+storiesOf('Scroller')
+	.addDecorator(withKnobs)
+	.addWithInfo(
+		'With ExpandableList',
+		() => (
+			<Scroller
+				style={{height: '600px'}}
+			>
+				<StatefullExpandableList
+					closeOnSelect
+					open={false}
+					title="Expandable List in Scroller"
+				>
+					{itemData}
+				</StatefullExpandableList>
+			</Scroller>
+		)
+	);