+++ packages/moonstone/CHANGELOG.md
@@ -46,6 +46,7 @@ The following is a curated list of changes in the Enact moonstone module, newest
 - `moonstone/Slider` and `moonstone/IncrementSlider` to not fire `onChange` for value changes from props
 - `spotlightDisabled` property support for spottable moonstone components
 - `moonstone/Popup` and `moonstone/ContextualPopupDecorator` so that when the popup is closed, spotlight focus returns to the control that had focus prior to the popup opening.
+- `moonstone/Scroller` to recalculate when an expandable child opens.
 
 ## [1.0.0-beta.1] - 2016-12-30